+++ Editor/Sources/Recorders/_Inputs/AOV/AOVCameraInputSettings.cs
@@ -6,6 +6,7 @@ namespace UnityEditor.Recorder.AOV.Input
     /// <summary>
     /// This class contains the information for an AOV.
     /// </summary>
+    [Obsolete("The AOV Image Sequence recorder is marked for deprecation and will be removed in Recorder 5.0.0.")]
     public class AOVCameraInputSettings : CameraInputSettings
     {
         /// <summary>