+++ Editor/Sources/Recorders/_Inputs/Camera/CameraInputSettingsPropertyDrawer.cs
@@ -73,8 +73,11 @@ namespace UnityEditor.Recorder.Input
                 EditorGUILayout.PropertyField(m_IncludeUI, Styles.IncludeUILabel);
             }
 
+#pragma warning disable 618
             if (!(target is AOVCameraInputSettings))
             {
+#pragma warning restore 618
+
                 // Hide the flip checkbox for AOV recorders
                 EditorGUILayout.PropertyField(m_FlipFinalOutput, Styles.FlipVerticalLabel);
             }